refactor: isolate sidebar footer presentation
