+++ Sources/SidebarDevFooter.swift
@@ -1,0 +1,30 @@
+#if DEBUG
+import CmuxFoundation
+import CmuxUpdater
+import SwiftUI
+
+/// Adds the development-build label to the shared sidebar footer controls.
+struct SidebarDevFooter: View {
+    var updateViewModel: UpdateStateModel
+    @ObservedObject var fileExplorerState: FileExplorerState
+    let modifierKeyMonitor: WindowScopedShortcutHintModifierMonitor
+    let onSendFeedback: () -> Void
+    let updatePillVisible: Bool
+    @AppStorage(DevBuildBannerDebugSettings.sidebarBannerVisibleKey)
+    private var showSidebarDevBuildBanner = DevBuildBannerDebugSettings.defaultShowSidebarBanner
+
+    var body: some View {
+        VStack(alignment: .leading, spacing: 6) {
+            SidebarFooterButtons(updateViewModel: updateViewModel, fileExplorerState: fileExplorerState, modifierKeyMonitor: modifierKeyMonitor, onSendFeedback: onSendFeedback, updatePillVisible: updatePillVisible)
+            if showSidebarDevBuildBanner {
+                Text(String(localized: "debug.devBuildBanner.title", defaultValue: "THIS IS A DEV BUILD"))
+                    .cmuxFont(size: 11, weight: .semibold)
+                    .foregroundColor(.red)
+            }
+        }
+        .padding(.leading, 6)
+        .padding(.trailing, 10)
+        .padding(.bottom, 6)
+    }
+}
+#endif
