+++ Sources/SidebarFooter.swift
@@ -1,0 +1,22 @@
+import CmuxUpdater
+import SwiftUI
+
+/// Keeps the footer's update visibility as a value read at the sidebar observation boundary.
+struct SidebarFooter: View {
+    var updateViewModel: UpdateStateModel
+    @ObservedObject var fileExplorerState: FileExplorerState
+    let modifierKeyMonitor: WindowScopedShortcutHintModifierMonitor
+    let onSendFeedback: () -> Void
+    let updatePillVisible: Bool
+
+    var body: some View {
+#if DEBUG
+        SidebarDevFooter(updateViewModel: updateViewModel, fileExplorerState: fileExplorerState, modifierKeyMonitor: modifierKeyMonitor, onSendFeedback: onSendFeedback, updatePillVisible: updatePillVisible)
+#else
+        SidebarFooterButtons(updateViewModel: updateViewModel, fileExplorerState: fileExplorerState, modifierKeyMonitor: modifierKeyMonitor, onSendFeedback: onSendFeedback, updatePillVisible: updatePillVisible)
+            .padding(.leading, 6)
+            .padding(.trailing, 10)
+            .padding(.bottom, 6)
+#endif
+    }
+}
